Update
Updated this version. Now allows the user to determine how the gradient works. original_form, random_selection, and weighted are all boolean types. If original_form = 1, then the algorithm performs the full gradient. If random_selection =1, then the algorithm samples the global push (currently the sample size is the same as the nearest neighbors size) and if weighted = 1, then it applies a weight to the distances corresponding the the outliers.